add css

+++ boardTest/WebContent/view.jsp
@@ -10,6 +10,7 @@
 <meta charset="UTF-8">
 <meta name="viewport" content="width=device-width" , initial-scale="1">
 <link rel="stylesheet" href="css/bootstrap.css">
+<link rel="stylesheet" href="css/custom.css">
 <title>JSP 게시판 웹 사이트</title>
 </head>
 <body>
@@ -123,7 +124,7 @@
 					if(userID !=null && userID.equals(bbs.getUserID())) {
 				%>
 						<a href="update.jsp?bbsID=<%= bbsID %>" class="btn btn-primary">수정</a>
-						<a href="deleteAction.jsp?bbsID=<%= bbsID %>" class="btn btn-primary">삭제</a>
+						<a onclick="return confirm('정말로 삭제하시겠습니다?')" href="deleteAction.jsp?bbsID=<%= bbsID %>" class="btn btn-primary">삭제</a>
 				<%
 					}
 				%>
